Replace old SQL with new SQL2.
